Extract selector logic into separate component
