+++ examples/create-react-app/src/App.js
@@ -20,7 +20,8 @@ import Legend from './components/legend';
 import PeriodPicker from './components/period-picker';
 import PeriodSelectorDialog from './components/period-selector-dialog';
 import PeriodSelector from './components/period-selector';
-import OrgUnitDialog from './components/org-unit-dialog'
+import OrgUnitDialog from './components/org-unit-dialog';
+import OrgUnitSelector from './components/org-unit-selector';
 import SelectField from './components/select-field';
 import Sidebar from './components/sidebar';
 import SvgIcon from './components/svg-icon';
@@ -159,6 +160,9 @@ class App extends Component {
 
                 <h2>Org unit dialog</h2>
                 <OrgUnitDialog d2={this.state.d2} />
+
+                <h2>Org unit selector</h2>
+                <OrgUnitSelector d2={this.state.d2} />
 
                 <div style={{ clear: 'both' }}>
                     <h2>TreeViews</h2>
