+++ packages/org-unit-dialog/src/index.js
@@ -1,3 +1,5 @@
 import OrgUnitDialog from './OrgUnitDialog';
 
+export { default as OrgUnitSelector } from './OrgUnitSelector';
+
 export default OrgUnitDialog;
